+++ src/commands/mod.rs
@@ -3,6 +3,7 @@
 
 // Messaging
 pub mod listen;
+pub mod listen_result;
 pub mod send;
 
 // Lifecycle